Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-mixed-content-to-inscope.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Not allowed to request resource
CONSOLE MESSAGE: Cannot load https://localhost:9443/service-workers/service-worker/resources/dummy?mode=no-cors&url=http%3A%2F%2Flocalhost%3A8800%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Not allowed to request resource
CONSOLE MESSAGE: Cannot load https://localhost:9443/service-workers/service-worker/resources/dummy?mode=no-cors&url=http%3A%2F%2F127.0.0.1%3A8800%2Fservice-workers%2Fservice-worker%2Fresources%2Ffetch-access-control.py%3FPNGIMAGE.


PASS Verify Mixed content of fetch() in a Service Worker 
